+++ src/pacman/model/GhostNodes.h
@@ -28,7 +28,7 @@ namespace PACMAN {
 
             const std::vector<Node*> get_respawns() const;
             const std::vector<Node*> get_spawns() const;
-            double get_cost(const Node& node) const;
+            float get_cost(const Node& node) const;
             const Node& get_node_towards_spawn(const Node& origin) const;
 
             const Node* get(int id) const;
@@ -45,7 +45,7 @@ namespace PACMAN {
             void add_respawn_paths();
 
         private:
-            std::map<const Node*, double> min_costs;  // best cost to reach said node
+            std::map<const Node*, float> min_costs;  // best cost to reach said node
             std::map<const Node*, const Node*> towards_spawn; // map node to a node closer to Pinky spawn
             std::vector<Node*> respawns;
             std::vector<Node*> spawns;
